Create README.md
The purpose of the article is represented by the automation of the process of drawing up the records of agricultural holdings through the necessary registers for the documentation and planning of plant protection works and the prevention of pollution with nitrates from agricultural sources. Through Robotic Process Automation (RPA) in agriculture, we can manage a wide range of repetitive or routine tasks. Using RPA to automate various agricultural operations allows companies and farmers to reduce unnecessary expenses while increasing production and profits. RPA is about simplifying complex agricultural processes that help save time and improve overall operational efficiency towards an important level of process planning and control that will enable farms to maximize their profitability with minimal losses. We carried out the Extraction, Transformation and Loading (ETL) of the data related to the declared parcels from the files downloaded from the holdings' account, later by means of the algorithms the summation of the areas related to the crops and categories of use found in the single payment request was realized. The sheets containing the data related to the animals declared by the holding and the identification data of the holding were identified, and finally the sheets intended for storing the extracted data and the calculated values were created.